services/api-server/pyproject.toml
[project]
name = "aibidder-api"
version = "0.1.0"
description = "AIBidder API service"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.14.0",
  "fastapi>=0.116.0",
  "uvicorn[standard]>=0.35.0",
  "pydantic-settings>=2.10.0",
  "sqlalchemy>=2.0.43",
  "python-jose[cryptography]>=3.5.0",
  "passlib[bcrypt]>=1.7.4",
  "python-multipart>=0.0.20",
  "psycopg[binary]>=3.2.10",
  "pypdf>=5.4.0",
  "httpx>=0.28.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4.2",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
